<refactor>: 메인화면 렌더링 방법 변경 및 전반적인 파일 위치 변경
