+++ package.json
@@ -9,6 +9,7 @@
     "polished": "^4.2.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-responsive": "^9.0.2",
     "react-router-dom": "^6.19.0",
     "react-scripts": "5.0.1",
     "react-social-login-buttons": "^3.9.1",
