+++ src/App.js
@@ -1,9 +1,10 @@
 import './App.css';
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import MainScreen from './screen/InitialScreen';
+import MainScreen from './screen/MainScreen';
 import LoginScreen from './screen/LoginScreen';
 // import GuideScreen from './screen/GuideScreen';
+import HomepageScreen from './screen/HomepageScreen';
 
 class App extends React.Component {
   render() {
@@ -13,6 +14,7 @@ class App extends React.Component {
           <Route path="/" element={<MainScreen />} />
           <Route path="/login" element={<LoginScreen />} />
           {/* <Route path="/guide" element={<GuideScreen />} /> */}
+          <Route path="/myhomepage" element={<HomepageScreen />}/>
         </Routes>
       </Router>
     );
