+++ src/component/main screen/AfterLogin.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import styled from 'styled-components';
+
+/* 사용자가 로그인한 후 화면 배치 */
+function AfterLoginSet() {
+
+}
+
+export default AfterLoginSet;
