+++ src/component/resource/naverLogo.png
